+++ DAL/ICommandesDB.cs
@@ -6,7 +6,7 @@ namespace DAL
 {
     public interface ICommandesDB
     {
-        int AddCommande(int idUtilisateur, int idLivreur, float prixTotal, int tempsLivraison, DateTime date);
+        int AddCommande(int idUtilisateur, int idLivreur, double prixTotal, int tempsLivraison, DateTime date);
         List<Commandes> GetCommandes();
         int UpdateCommandeLivree(int idCommande);
     }
