+++ DAL/IPlatsDB.cs
@@ -7,6 +7,6 @@ namespace DAL
     {
         int GetPlatID(string nom, int idRestaurant);
         List<Plats> GetPlats();
-        float GetPrixPlat(int idPlat);
+        double GetPrixPlat(int idPlat);
     }
 }
